IoUring: Use correct native jni types in function definition
Motivation:

When define functions that are exposed via JNI we need to ensure we use the types that are provided and supported by JNI. Not doing so might result in undefined behaviour depending on the platform.

Modification:

Fix netty_io_uring_unregister_buf_ring to use the correct supported types.

Result:

No undefined behaviour anymore when unregister buffer ring

+++ transport-classes-io_uring/src/main/java/io/netty/channel/uring/Native.java
@@ -485,7 +485,7 @@ final class Native {
     static native int ioUringRegisterRingFds(int ringFds);
 
     static native long ioUringRegisterBufRing(int ringFd, int entries, short bufferGroup, int flags);
-    static native int ioUringUnRegisterBufRing(int ringFd, long ioUringBufRingAddr, int entries, int bufferGroupId);
+    static native int ioUringUnRegisterBufRing(int ringFd, long ioUringBufRingAddr, int entries, short bufferGroupId);
 
     static native int ioUringEnter(int ringFd, int toSubmit, int minComplete, int flags);
